Fixes to welcome page content + removal of plans page
