+++ public/index.html
@@ -1,18 +1,8 @@
 <head>
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <link rel="shortcut icon" href="img/favicon.ico" type="image/x-icon" />
+  <!-- <link rel="shortcut icon" href="img/favicon.ico" type="image/x-icon" /> -->
   <link rel="apple-touch-icon" href="img/apple-touch-icon.png" />
-  <link
-    rel="apple-touch-icon"
-    sizes="72x72"
-    href="img/apple-touch-icon-72x72.png"
-  />
-  <link
-    rel="apple-touch-icon"
-    sizes="114x114"
-    href="img/apple-touch-icon-114x114.png"
-  />
 
   <link rel="stylesheet" type="text/css" href="css/bootstrap.css" />
   <link
@@ -39,7 +29,7 @@
     href="https://fonts.googleapis.com/css?family=Raleway:300,400,500,600,700,800,900"
     rel="stylesheet"
   />
-  <title>A React Landing Page Template</title>
+  <title>ChurroFlow</title>
   <meta name="description" content="" />
   <meta name="author" content="@Issaafalkattan" />
 </head>
